+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -23,6 +23,7 @@
 - Added experimental `distribution: ray` support in command job.
 - Added support to enable data isolation feature at workspace creation stage.
 - Added auto_delete_setting support for asset version in data import job.
+- Added support to enable set workspace connection secret expiry time.
 
 ### Bugs Fixed